refactor: cookie-consent tweaks and renamed utils
