+++ components/cookies/_script.tsx
