+++ components/cookies/banner.tsx
@@ -1,22 +1,11 @@
+"use client";
+
 import Link from "next/link";
 import { Slot } from "@radix-ui/react-slot";
 import { Cookie } from "lucide-react";
 
 import { cn } from "@/lib/utils";
-import { Button, type ButtonProps } from "../ui/button";
-
-import { GroupedOptions, ShowMeButton } from "./banner-options";
-import { getCookie } from "cookies-next";
-import { ANALYTICS_TAGS, NECESSARY_TAGS, tagDetails } from "./utils";
-
-type BannerProps = React.PropsWithChildren<
-  {
-    bannerClass?: string;
-    asChild?: boolean;
-    buttonGroup?: React.ReactNode;
-    leftElement?: React.ReactNode;
-  } & BannerContentProps
->;
+import { BannerTriggerGroup } from "./banner-buttons";
 
 export const background =
   "bg-muted/20 py-4 px-6 rounded-lg shadow-lg flex items-center justify-between gap-x-4 backdrop-blur-md";
@@ -39,13 +28,6 @@ export default function Banner(props: BannerProps) {
   );
 }
 
-type BannerContentProps = React.PropsWithChildren<{
-  heading?: string;
-  description?: string;
-  href?: string;
-  label?: string;
-}>;
-
 function BannerContent(props: BannerContentProps) {
   return (
     <div className="flex flex-col justify-center gap-y-2 mr-2 flex-1 text-sm">
@@ -61,104 +43,6 @@ function BannerContent(props: BannerContentProps) {
         </Link>
         .
       </p>
-    </div>
-  );
-}
-
-type ButtonGroupProps = React.PropsWithChildren<
-  {
-    asChild?: boolean;
-  } & BannerTriggersProps
->;
-
-export function BannerTriggerGroup(props: ButtonGroupProps) {
-  const ButtonGroupSlot = props.asChild ? Slot : BannerTriggers;
-  return (
-    <div className="flex flex-col md:flex-row gap-y-2 md:gap-x-2">
-      <ButtonGroupSlot {...props}>{props.children}</ButtonGroupSlot>
-    </div>
-  );
-}
-
-type BannerTriggersProps = {
-  buttons?: ButtonProps[];
-  asChild?: boolean;
-};
-
-const _buttons: BannerTriggersProps["buttons"] = [
-  { children: "Show Me", variant: "outline", type: "button", size: "sm" },
-  { children: "Got it", variant: "default", type: "submit", size: "sm" },
-];
-
-const isPro = !(process.env.NEXT_PUBLIC_FEATURE_PRO === "true");
-
-function BannerTriggers(props: React.PropsWithChildren<BannerTriggersProps>) {
-  const { asChild, buttons, children, ...rest } = props;
-
-  let btns = buttons ?? _buttons;
-  if (btns && btns.length > 2) {
-    btns.length = 2; // removes all buttons after the 2nd
-    console.log(btns);
-    console.warn("BannerTriggers: Only 2 buttons are supported");
-  }
-
-  return asChild ? (
-    <Slot>{children}</Slot>
-  ) : (
-    <>
-      {btns
-        ? btns.map((btn, i) => {
-            if (isPro && i === 0) {
-              // only show the feature button if the user has pro subscription
-              return <ShowMeButton key={i} {...btn} />;
-            }
-            return <Button key={i} {...btn} {...rest} />;
-          })
-        : null}
-    </>
-  );
-}
-
-export function BannerOptions({
-  tags,
-}: {
-  tags: NecessaryTrackingTagsTupleArrays;
-}) {
-  const titles = ["Necessary", "Analytics"]; // @TODO: add support for 3rd party tags
-  const cookies = JSON.parse(
-    getCookie("app-consent") || "{}"
-  ) as BrowserCookies;
-
-  const options = {
-    ...NECESSARY_TAGS.reduce((acc: any, tag: string) => {
-      acc[tag as keyof typeof acc] = {
-        ...tagDetails[tag as keyof typeof tagDetails],
-        checked: cookies[tag as keyof typeof cookies],
-      };
-      return acc as AllOptions;
-    }, {}),
-    ...ANALYTICS_TAGS.reduce((acc: any, tag: string) => {
-      acc[tag as keyof typeof acc] = {
-        ...tagDetails[tag as keyof typeof tagDetails],
-        checked: cookies[tag as keyof typeof cookies],
-      };
-      return acc as AllOptions;
-    }, {}),
-  };
-  return (
-    <div className="grid gap-4 min-w-2xl p-2 bg-background/40 backdrop-blur-md rounded-md z-10">
-      {tags.map((tagGroup, index) => {
-        return (
-          <GroupedOptions
-            key={index}
-            tagGroup={tagGroup}
-            options={options}
-            description=""
-            tags={tags}
-            category={titles[index]}
-          />
-        );
-      })}
     </div>
   );
 }
