+++ components/cookies/utils/constants.ts
@@ -1,0 +1,51 @@
+export const CONSENT_COOKIE_NAME = "app-consent";
+export const DATA_LAYER = "dataLayer";
+export const TAG_MANAGER_KEY = "gtag";
+
+export const NECESSARY_TAGS = [
+  "security_storage",
+  "functionality_storage",
+  "personalization_storage",
+];
+
+export const ANALYTICS_TAGS = [
+  "ad_storage",
+  "analytics_storage",
+  "ad_personalization",
+  "ad_user_data",
+];
+
+export const tagDetails: TagDetails = {
+  security_storage: {
+    label: "Security Related Cookies",
+    description: "Cookies necessary for securely authenticating users.",
+  },
+  functionality_storage: {
+    label: "Functionality Related Cookies",
+    description: "Cookies for measuring and improving site performance.",
+  },
+  personalization_storage: {
+    label: "Personalization Related Cookies",
+    description: "Cookies for enhanced functionality and personalization.",
+  },
+  ad_storage: {
+    label: "Personalized Marketing Related Cookies",
+    description: "Cookies for targeted content delivery based on interests.",
+  },
+  analytics_storage: {
+    label: "Analytics Related Cookies",
+    description: "Cookies for measuring and improving site performance.",
+  },
+  ad_personalization: {
+    label: "Personalization Related Cookies",
+    description: "Cookies for enhanced functionality and personalization.",
+  },
+  ad_user_data: {
+    label: "User Data Related Cookies",
+    description: "Cookies for targeted content delivery based on interests.",
+  },
+};
+
+export const redactionCookie = "ads_data_redaction";
+
+export const cookieExpiry = 60 * 60 * 24 * 7; // Set expiration (1 week)
